backend/pyproject.toml
[tool.poetry]
name = "tabby-sync"
version = "1.0.0"
description = "Minimal Tabby config sync backend"
authors = []

[tool.poetry.dependencies]
python = "^3.14"
Django = "^5.1"
djangorestframework = "^3.15"
python-dotenv = "^1.0"
dj-database-url = "^2.2"
gunicorn = "^23.0"
django-cors-headers = "^4.4"
whitenoise = "^6.7"
mysqlclient = "^2.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
